fixed missing data issue in allgatherv

+++ cpp/src/cylon/net/ucc/ucc_operations.hpp
@@ -28,8 +28,8 @@ class UccTableAllgatherImpl : public net::TableAllgatherImpl {
   ucc_context_h ucc_context_;
   std::vector<ucc_coll_req_h> requests_;
   std::vector<ucc_coll_args_t> args_;
-  std::vector<std::vector<uint64_t>> counts_;
-  std::vector<std::vector<uint64_t>> displacements_;
+  std::vector<std::vector<uint32_t>> counts_;
+  std::vector<std::vector<uint32_t>> displacements_;
   int rank;
   int world_size;
 };
@@ -70,8 +70,8 @@ public:
  private:
   std::vector<ucc_coll_req_h> requests_;
   std::vector<ucc_coll_args_t> args_;
-  std::vector<std::vector<uint64_t>> counts_;
-  std::vector<std::vector<uint64_t>> displacements_;
+  std::vector<std::vector<uint32_t>> counts_;
+  std::vector<std::vector<uint32_t>> displacements_;
   ucc_team_h ucc_team_;
   ucc_context_h ucc_context_;
   int world_size;
